mark elements imported via `require class` as synthetic with shallow class decl
Summary:
The test `multiple_01.good.php` was accepted with folded decls but it was rejected with shallow class decl. This can be explained via the logic in `Typing_extends.check_trait_diamond`: synthetic elements should not take part in diamond checks.

This diff ensures that elements imported via `require class` trait constraints are marked as synthetic.

Reviewed By: vsiles

Differential Revision: D35360786

fbshipit-source-id: bba77982fa37e9c54c1a613482b0597b1f79c6e8