Stop sub-type logging from printing Capabilities
Summary: `hh_log_level('sub', 2)` is filled with capability related calls that aren't in the code and often obscure what I'm trying to debug. This turns them off unless the level is 3 or higher, since they are not useful except when trying to debug actual capability issues. Does this seem like something we should land?

Reviewed By: madgen

Differential Revision: D35361737

fbshipit-source-id: eeb1cd4438262e28c01cc88b90cc5e566e3ba370