hhbc/ast_constant_folder.rs -> emitter/constant_folder.rs (4/N)
Summary:
Like the rest of `emitter`, constant_folder transforms AST into bytecode and
does not define data types that are part of HackCUnit. It belongs in hackc/emitter.

Ideally, there should only be one place in the compiler that implements the
semantics of Hack arithmetic; currently both constant_folder and typed_value.rs
contain constant-folding logic. I started to consolidate it but decided to split the
rest into a separate diff (next).

Reviewed By: aorenste

Differential Revision: D35340760

fbshipit-source-id: 043b16ffd03e826d6320d519f3aa574b6e7a89ac