HhasSymbolRefs should not depend on emitter SymbolRefState (6/N)
Summary:
We use SymbolRefState to juggle symbol refs during compilation (using BTreeSets)
and then linearize the final results into a HhasSymbolRefs in `hhbc`.  Move the final
linearization step to emitter so we don't have a backwards dependency.

`IncludePath` *is* part of HackCUnit and so it belongs in `hhbc`, not emitter. Moved it.

After these changes, symbol_refs_state.rs was such a small module that it doesn't justify
being in its own crate. Merged it into the `env` crate.

Reviewed By: aorenste

Differential Revision: D35342084

fbshipit-source-id: 26d50915af38e44cfd9c76837074900c6344152f