Move Label -> hhbc_ast, LabelGen -> env (7/N)
Summary:
`Label` is part of the `Instruct` ADT and referenced by `HhasParam`, but
`LabelGen` is belongs in emitter, now as part of the `env` crate.

`Local` and `IterId` will get similar treatment, next.

`InstrSeq::create_try_catch()` moved to `scope.rs` because it uses LabelGen,
and because that functionality is closer related to scope.rs anyway.

Reviewed By: aorenste

Differential Revision: D35342086

fbshipit-source-id: f16e42f20c9f83180b1755a8d73dbe58e73bd9ae