TypeConstraint: do not conflate object and unresolved types
Summary:
Previously, AnnotType::Object meant an unresolved name rather than an object.
Except of an object, this name could have been a type alias name or an enum
name. We tried to resolve these names in hhbbc and updated the type if the
result was a non-object. However, if the resolved type was an object, we could
not update it, because it would break error messages. We just marked it
as resolved, but in that case the only piece of information we could use is
that it was an object. We still had to resolve the class name again at runtime,
which was not always possible or practical.

Fix this by starting at AnnotType::Unresolved and adding a separate field for a
resolved class name if we resolve to AnnotType::Object.

Audit all uses of AnnotType and TypeConstaint and take advantage of this new
information.

Reviewed By: ricklavoie

Differential Revision: D35394205

fbshipit-source-id: 9bf654f4a3c58b4cee1a99a492686e79da471b3a