Include StringishObject ancestors
Summary:
Previously, we were only adding `StringishObject` to the list of ancestors. We were not correctly recursing over the ancestors of `StringishObject` and inheriting the ancestors. In order to do so, we need to recurse over `StringishObject` but we also need to include `StringishObject` in the list of `decl_class_parents` for a class.

OCaml updated the `decl_class_parents`, but in Rust, it appears that we do this all at once, so I've added the logic to check whether StringishObject is a valid parent there as well.

Differential Revision: D35374704

fbshipit-source-id: af7f40821d40950bc22b901ecf04f20694719096