Avoid using prop_to_env outside subtyping module
Summary:
Now that subtyping clients are responsible for managing an error "result", we don't need to use `prop_to_env` outside the subtyping module. Its one use is in determining the element type of a container when doing memoize signature checks.

We should be able to restrict `simplify_subtype_i` similarly, except for its single use in `Typing_ordered_solver`.

Reviewed By: vsiles

Differential Revision: D35408021

fbshipit-source-id: ef1695e722918fed7f57bde03c920a39dfe1bb59