Thread DeclParserOptions into DeclParser in compare_folded_decls_file
Summary:
In the single-file comparator test runner, we collect options from CLI arguments in HH_FLAGS or .flags files. For OCaml decl, those options are fed into the direct decl parser. But for the Rust implementation, we haven't yet threaded those options through.

In order to avoid adding a lifetime to `DeclParser` (and every other struct containing one), we need it to be able to own its DeclParserOptions, but the only version we have is the one in oxidized_by_ref. We could make an owned version (either by hand-writing it or generating it from declParserOptions.ml in oxidized), but it would require a conversion to the by-ref version every time we would like to parse a file. Instead, this diff adds a wrapper struct `hackrs::decl_parser::options::Options` which contains both a `DeclParserOptions` and the arena in which it is allocated (using the ouroboros library for self-referential structs).

Reviewed By: edwinsmith, shayne-fletcher

Differential Revision: D35450631

fbshipit-source-id: 8f1098ef573ba4f7b82b2c72c6d706f5ef410245