Fix hhbc-ast.h generation to include the macro-generated Opcode enum
Summary:
The Opcode enum is filled in by a macro which is invisible to the ffi_cbindgen program.  So instead we have to use dump-opcodes to emit the output of the macro and pass that file to ffi_cbindgen instead.

dump-opcodes tried to be clever and pipe its output through rustfmt - unfortunately since it just calls "rustfmt" directly it didn't handle the case where our build machines didn't have rustfmt in the path.

While I was adding command-line option support to dump-opcodes anyway I added an `--output` flag.

Reviewed By: kavoor

Differential Revision: D35453949

fbshipit-source-id: 7f5206f57143358dcd099b08b492a1812556df7d