Add new internal field to class definition in AST
Summary:
This stack of diffs will address 1. adding a field to represent the "internal" modifier to all relevant places in the AST where it can eventually be used (e.g. functions, classes, traits, interfaces) on the oCaml side of things 2. updating the Rust view of the AST for parsing and 3. removing the internal field from file attributes. This will all allow us to use the internal keyword directly instead of as part of an attribute (i.e. `<<__Internal>>`)

This diff addresses part 1. for classes.

Reviewed By: jamesjwu

Differential Revision: D35264742

fbshipit-source-id: d8c5ea0ce56f2c7bf4bfbc2338ea3c22f2c5bafc