Emit at most one fallback/fail block per type verification
Summary:
If the type constraint is unresolved, do not emit one fallback per each
possible input data type. Instead, keep track of the most pessimistic needed
fallback mode (always fail < fallback < fallback with coercion) and emit it
once if none of the better cases were handled.

Note that this change does nothing yet for parameter and return types, as the
fallback is currently implemented by PUNT(). However, it allows the PUNT() to
be replaced with a proper implementation without causing a code bloat. That
will come in the next diff.

This change doesn't do much for property types either, as multi-datatype inputs
always use the fallback. It prepares the code for the future improvements
though.

Reviewed By: mofarrell

Differential Revision: D35451504

fbshipit-source-id: 0936f011239f49b98331b57c00d060d1e155c8ef