Support multiple constraints in type const at the syntax level
Summary:
Currently, the syntax node that corresponds to `const type`
associates only a single type constraint that is upper bound (`as`).
In order to support multiple constraint and/or lower bounds (`super`),
change the representation to be a list of constraints.

The lowerer can then partition the bounds into lower and upper sets,
computing union type of the former and intersection type for the latter. E.g.
```
class A { const type T = num; }
abstract class C<T as I> {
  abstract const type T as arraykey as A::T; // now supported by parser
  abstract const type T2 as arraykey super int super A::T;
}  // NOTE: `super` is not yet supported but it will be at some point
```
should lower into `T as (arraykey & A::T)` and `T2 as arraykey super (int | A::T)`.

For backward compatibility, ban this in the typechecker (unless opted-in from tests using `--allow-unstable-features`),
but allow them in the HackC/HHVM parser (so we don't slow the rollout later)
because bounds on type constants are invisible to the HHVM anyway.

Currently, the parser only supports upper bounds;
the idea is to allow lower bounds (`super`) via another experimental feature.

Reviewed By: vassilmladenov

Differential Revision: D35202390

fbshipit-source-id: f190a2b1be7594138d73b55d8acb965fb792999e