Match OCaml behavior for unwrap_class_type
Summary:
In OCaml, `unwrap_class_type` has some default behavior for non-Tapplys:

https://www.internalfb.com/code/fbsource/[33fad67722ae]/fbcode/hphp/hack/src/decl/decl_utils.ml?lines=33-41

But the Rust implementation only handles the Tapply case.

While every ty which we invoke unwrap_class_type on will be a Tapply in a correct program, we'd like to behave the same as the OCaml for incorrect programs too.

Reviewed By: shayne-fletcher

Differential Revision: D35479751

fbshipit-source-id: 3402bba49d36f13bdc4a9c2b11f1c3de77029866