Use supportdyn for methods
Summary:
Wrap the type of a method from a `__SupportDynamicType` class (or if the method itself has the annotation) in `supportdyn<method_type>` before checking a call to it. This follows how we check static methods and top-level functions. There were a number of places where code was inspecting types and looking for `Tfun _`, which needed to also detect `Tnewtype ("supportdyn", TFun _, _)`.

Using the supportdyn rule introduces a soundness bug with `__RequireDynamic`, but that feature is already unsound, and needs improvement anyway. See the `test/sound_dynamic/typing/soundness_bugs/generic_params_2.good.php` for an example.

Reviewed By: vassilmladenov

Differential Revision: D35363412

fbshipit-source-id: 8455eb12ca3286c835b14e5bc72741c48655ecc7