Remove AstBody
Summary:
I've been going through the modules in the `aast` directory (crate) and doing
something about the ones that seem out of place.

AstBody is a utility enum that either wraps a vec of Defs or Statements; it is used
in one place to make `compute_vars()` handle both cases, but everywhere else
it only is used for statements. Whatever it's original goals were, at this point it doesn't
seem to be providing enough utility to justify it's existence.

In the one place we need it, `compute_vars()` can take `&impl Node`; everywhere
else we can just use `&[Stmt]` or `&[Def]`.

Reap dead emit_body() code revealed by removing the impossible AstBody::Defs case.

Reviewed By: aorenste

Differential Revision: D35348458

fbshipit-source-id: ddb5d3edeb261699b6db6f664456b231b3f98cf0