Pass C/C++ compiler to dune (#9065)
Summary:
As discussed in https://github.com/hhvm/packaging/pull/287 , there is a bug in g++ 11.2 preventing folly and thrift from being compiled. We will compile hhvm on 21.10 with gcc 10 instead. Therefore, this PR is used to pass the compiler to subprojects.

In https://github.com/facebook/hhvm/issues/9048 and https://github.com/facebook/hhvm/pull/9052, we have added the C and C++ compiler to all third-party libraries. This diff in addition passes the compiler setting to `Makefile.dune` as well.

Note that we might still need an additional diff for `Makefile.dune` to pass the compiler to `dune build` command.

Pull Request resolved: https://github.com/facebook/hhvm/pull/9065

Test Plan: This diff should not break `hhvm-and-hack-ubuntu` and `hhvm-and-hack-mac` Sandcastle jobs

Reviewed By: fredemmott

Differential Revision: D35502716

Pulled By: Atry

fbshipit-source-id: 8b241d336372536baf2b89b54d343b3ea24fc827