Add `//hphp/system:typecheck`
Summary:
**Context:** `systemlib` comprises all of the PHP code that is shipped with HHVM and executed but never checked by Hack. This code ought to, and might *have* to, receive the same level of scrutiny we give to any userland piece of Hack code.

This diff creates a test that gathers a subset of PHP files in `hphp/system/` and checks to see whether they typecheck together according to `hh_single_type_check`. This is created with a block-list for files that we do not want to include. As new files are added, they will default to being typechecked.

We currently typecheck a single file: `hphp/system/php/lang/ArrayAccess.php`, which I've added generics to, to match it's HHI definition.

Reviewed By: aorenste

Differential Revision: D35404099

fbshipit-source-id: f3164305722db39daa8bc43fbd8f028046ab3f08