Remove global RAT array table
Summary:
Similar to the literal string and literal array tables, remove the
global RAT array table and replace it with per-unit tables. While I'm
at it, add support for lazy loading of the RAT arrays as well.

There's no need for the RepoAuthType logic to be split into 3
different sets of files/headers, so consolidate it all into one.

Reviewed By: edwinsmith

Differential Revision: D35457052

fbshipit-source-id: 30a6b0903f5ee334a0256eb03b87bb1b46e0a6ca