Remove the global literal array and string tables
Summary:
If we wish to build a repo-file in a more distributed way, it becomes
difficult to support single global literal array and string tables. We
already have support for per-unit literal and string tables (this is
needed for Sandboxes), so just always use those. This change actually
tends to produce more compact bytecode, since the literal ids are, on
average, much smaller now.

We did not support lazy loading of strings and arrays for the per-unit
tables, so add support for that. We do no longer need the "use global
ids" flag for BlobDecoder or UnitEmitter, so that can be removed.

Reviewed By: edwinsmith

Differential Revision: D35456802

fbshipit-source-id: 6d802ccdbfe45d76b2db1f1ad7911a6f7f877808