Tighten and de-duplicate some lvalue checks
Summary:
**Context:** I'm trying to change Hack to treat lvals more consistently, with the ultimate goal of encoding lvals distinctly within Hack itself.

There are a few different places where we check for "lvalues" in Hack, including but not limited to:
* `RustParserErrors::check_lvalue`
* `lowerer::check_lvalue`
* One-off rules in the typechecker like `Call_in_subscript_check` and the `Primary.Wellformedness.Dollardollar_lvalue` error

This doesn't cover that we have a separate pass for checking `inout` arguments, which have slightly stricter rules than lvalues (e.g. disallowing class member selection).

The primary change here is that `RustParserErrors::check_lvalue` now defaults to erroring: you need to, by hand, teach Hack that a given value is a valid lvalue. This allows us to:
* Totally remove `lowerer::check_lvalue`, as it's now completely covered by this check. Before it haphazardly covered some scenarios like `foo() = 42`.
* Totally remove some of the aforementioned typechecker rules as they're now covered by the parse-time check.

In a follow up I'll probably unify some of the `inout` errors as well.

Reviewed By: Wilfred

Differential Revision: D35054000

fbshipit-source-id: 29d4417c95f6078faf26c690c038a1bb4655b01c