Unlock lazy load locks on exceptions
Summary:
There's situations where we can throw exceptions while
loading data out of the Repo (for example, if we invoke the
unserializer). If that happens, the lock will be kept lock and
anything trying to read it after that will deadlock.

Use SCOPE_FAIL to make sure we unlock on the way out.

Reviewed By: paulbiss, mofarrell

Differential Revision: D35563589

fbshipit-source-id: c1df0de792b125a954a29abab76beba302fe9974