use mini-state in indexer
Summary:
Currently, glean indexer always does a full-init, regardless of the size of the shard it indexes. We want to be able to use a saved state to save some parsing time in those workflows where we only index a few files.

A typical invocation of `hh_server` for indexation is:

`hh_server --write_symbol_info DIR  --config use_mini_state=true --config symbol_write_index_paths_file=FILES_TO_INDEX`

This does a full init. This diffs adds  two new ways to invoke the indexer:

1. `hh_server --write_symbol_info DIR  --config with_mini_state=... `. It gets a precomputed state from the file system.

2. `hh_server --write_symbol_info DIR  --config use_mini_state_when_indexing=true ...`. Download the state from manifold.

The use case I have in mind is to use 2 in our skycastle workflows.

Note that the behavior of the option `use_mini_state=true` is incorrect when used in conjunction with  `--write_symbol_info` : it's mandatory, regardless of what we want to do with the saved state). To circumvent this, I added a new option `with_mini_state_when_indexing`. Didn't see how to reuse existing options, with the constraint of not breaking the previous behavior .

Reviewed By: ljw1004

Differential Revision: D35062485

fbshipit-source-id: b1aefa1ba635f4afd3443ef23286490de49f8936