Throw if we can't query Watchman
Summary: Today, if native Facts can't initialize or query Watchman, we just proceed with stale Facts. We'll return incorrect answers until we're able to successfully query Watchman and update Facts again. This is bad. Instead, we should throw. With the current behavior, this will cause HHVM to revert to userland autoloading via SQLFacts. This is fine for now.

Reviewed By: periodic1236

Differential Revision: D35560093

fbshipit-source-id: ef6795425aaa785b1b8785f66ec92ea677eebdea