Telemetry: remove is_lazy_incremental
Summary:
I don't trust how is_lazy_incremental is handled. It's a global reference that sometimes gets set with a setter. Looking at code around that reference in HackEventLogger, it seems non-trivial to get it right, and impossible to convince oneself it is correct.

Besides, for some events like recheck_end event, for example, it's completely meaningless because a "recheck" is a loop with multiple checks per iterations, where each check can either be lazy or full. We now have a check_kind field which is properly set with a with_check_kind helper, so is_lazy_incremental is redundant.

Reviewed By: ljw1004

Differential Revision: D35547976

fbshipit-source-id: a9c0f23e19f1ed9029ddc36c7c93fc0cdc83952e