Use an explicit stack for PostorderSort
Summary:
PostorderSort makes use of the call stack for its DFS walk of the unit.  Create
a heap allocated stack for the walk so that we don't blow the stack, and get
the appropriate max size of 8 bytes per block.  Rather than incur resizes this
allocates enough space for all blocks to be on the stack simultaneously.

Reviewed By: ricklavoie

Differential Revision: D35561888

fbshipit-source-id: e48781a6739a89fc58198439a08552c8fa434463