Telemetry: rename rechecked_count to changed_files_count and others
Summary:
The Scuba column rechecked_count for event RECHECK_END is confusing and wrong. This is actually the number of files changed, either on disk or in the IDE. This is called "dirty" files somewhere else in the code base, although "changed" is more explicit? I don't have a strong opinion.

So I'm changing the name in most places.
I've made sure to keep the old Scuba column as well to avoid breaking dashboards.

Reviewed By: andrewjkennedy

Differential Revision: D35549549

fbshipit-source-id: de01bd55938efc4b636391c9d768d475b2a579de