typing constraints for foreach (boilerplate code)
Summary: Add a new constraint type that represents the ability to traverse an array with `foreach`. This will let us move the checking logic into the core subtyping algorithm. We can then reduce the amount of redundant code for handling unions, intersections, and dynamic, while improving the treatment of generics. This is just the boilerplate to get the new constraint type into the code, it should make no visible changes.

Reviewed By: enetsee

Differential Revision: D35579392

fbshipit-source-id: 5e66e53c66a7e47fd3857e5cfc184e677a9be6a9