Change alignment of ObjectData to eliminate 4 bytes between it and props
Summary:
There are four wasted bytes between an `ObjectData` and its associated `ObjectProps` when 1) we're using low pointers and 2) compiling with MSVC.

We can change the alignment on the `ObjectData` struct so that `ObjectData::props()`'s `(this + 1)` returns this unaligned address (https://www.internalfb.com/code/fbsource/fbcode/hphp/runtime/base/object-data-inl.h?lines=404).
In addition, since we don't support compiling with MSVC, that check has been removed as a condition for packing the struct.

Reviewed By: mofarrell

Differential Revision: D35015868

fbshipit-source-id: ed5772975ba60401f85c047bb74840b9e4981820