move shmffi::HeapValue into new crate ocaml_blob
Summary:
The structure "HeapValue" represents an ocaml marshalled+compressed blob which gets stored in the hashtable. (Well, if it's a string then it's not marshalled, and if compression enlarges it then we skip compression).

At the moment, HeapValue is inside shmffi, the vehicle by which sharedmem.ml and hh_shared.c access the sharded hashtable. But I want hh_decl_shm to be able to store the exact same ocaml marshalled+compressed blobs as well. Therefore, I'm pulling it out into a new crate.

(I kind of thing HeapValue is too generic a name. We should rename it as OcamlSerializedCompressedBlob. That's because other usecases will store other kinds o HeapValue on the heap. But I'll leave this rename pending Henri's approval.)

Differential Revision: D35514191

fbshipit-source-id: fea70755aa29fb182cb89a335a15c54ff414f21c