move shmffi::ShmemSegment into shmrs::
Summary:
The ShmSegment structure represents a shared-memory segment that stores a sharded hashtable.

It used to be inside module "shmffi", the vehicle by which shared_mem.ml and hh_shared.c can access sharded hashtable.

But I want the Rust hh_decl_shm to be able to access the sharded hashtable also. Therefore, I'm moving ShmSegment out of shmffi. I decided to move it directly into shmrs, the crate which deals with the implementation of the sharded hashtable. I think the two go hand in hand.

Previously, ShmSegment was hard-coded to HeapValue (an ocaml marshalled+compressed blob). I've changed it over to a generic TValue, so it can be used with any format blob we chose.

Differential Revision: D35485300

fbshipit-source-id: 178da8110dac009e512fdb0f31ff154ea12a74ad