Translate HhasAdata and delete adataDecls
Summary:
Translate HhasAdata

Delete `adataDecls`. This map was added in D14881020 (https://github.com/facebook/hhvm/commit/1e80933ffa2d0a4fa54ba89a67cb20e3d5f84263) because we needed to add a provenance tag to the array before we merged it. This provenance tag required the srcloc, something we didn't have until we saw the array on an `IMM_AA`. I believe this is no longer a dependency, and we should be able to unserialize and merge the array from the `.adata` again.

Not implementing `toTypedValue()` for HhasAdata, as `__hhasadata()` is being deprecated.

Reviewed By: ricklavoie

Differential Revision: D35541008

fbshipit-source-id: f3650f97dea0afcb8ea106cb31935d481c029fb9