typing constraints for foreach
Summary:
Switch typechecking of `foreach` to use the `Tcan_traverse` constraint type.

This also fixes a bug where in a `foreach ($x as $v)` or `foreach ($x as $k => $v)` if `$x : dynamic & C` where C was an appropriate implementation of `Traversable` or `KeyedTraversable`, then `$k` and `$v` got typed as `dynamic & C`, instead of `dynamic & t` where `t` is the key/value type from `C`.

Reviewed By: enetsee

Differential Revision: D35551326

fbshipit-source-id: 3df4267f963840a957b5e3b809f48cca592bc4e4