Prepare systemlib and runtime tests for enforcement of reading globals needing
Summary:
D33949450 (https://github.com/facebook/hhvm/commit/7bd1acb099206d127b89a869049e7d284d95419c) was reverted due to causing a SEV.

This happened when a caller of `Exception::getTraceOptions()` in www did not
expect a readonly return value even though it was updated to be readonly int in
the systemlib defintion, causing mismatch at runtime.

To re-land the original changes, it is safer to split them into different
diffs (to reduce surface area of the changes). This diff can be viewed as a
preparation for the actual runtime enforcement behavior changes. It modifies
all the places where readonly is expected (e.g. within runtime tests and
`Exception::getTraceOptions()`). The next diff adds the enforcement back.

Reviewed By: oulgen

Differential Revision: D35558771

fbshipit-source-id: 087c55c867c3c19938eedcc68aa276c9ac2c55a0