Fix iteration order for stringish object
Summary:
There is a subtle order dependent behavior in the way that we inherit ancestors.
This particularly affects `StringishObject`.
Currently, when we go to inherit `StringishObject`, if we inherit `StringishObject` from an ancestor, we will get the wrong position for the reason and the type. Instead of getting the position of the `__toString` method of the current class, we will get the position of the `__toString` method of the class that we iterated over.

The fix to match the OCaml code is to always inherit the `StringishObject` of the current class _last_. That way, we overwrite any of the parent's versions of `StringishObject` and replace them with the one with the current class's `__toString` position.

In OCaml we do the following, inserting the new `StringishObject` at the front of the list of implements rather than at the end.

https://www.internalfb.com/code/fbsource/[3e71742608d0]/fbcode/hphp/hack/src/decl/decl_folded_class.ml?lines=858

Differential Revision: D35600004

fbshipit-source-id: cdd3be8ee783423ffa1d8cbf61bb7dbbfaa036da