Convert `watchman.socket.default` into a template string with an expandable "%{user}" tag
Summary:
We use Watchman to track the files and file changes on a given repository. To communicate with the correct Watchman server, we first check which user owns the repo we're watching, then we resolve that username to a socket path.

Today, we "resolve that username to a socket path" by finding the user's homedir, finding a `.hphp` file in the homedir, and finding a `watchman.socket.default` option in the `.hphp` file. This is kind of a roundabout way to do it. Also, not all users have homedirs with `.hphp` files. Fortunately, our watchman sockets are all in paths of the form `/var/facebook/watchman/$USER-state/sock`. So instead of going through homedirs, we can just write a template string: `"/var/facebook/watchman/%{user}-state/sock"`. Then we can expand `%{user}` to the correct username.

Reviewed By: periodic1236

Differential Revision: D35412413

fbshipit-source-id: e65f01c693db26cf2b695454a172b3b2fadf6c69