implement AsyncMysqlResult::sslSessionReused() method
Summary: Currently `sslSessionReused()` is implemented only in `AsyncMysqlConnection` class: https://fburl.com/code/mgyrdbvr. The connection object is available at the points when we do logging only in scenarios where we explicitly request a connection object to be created, and where we succeed in it. In situations where connection failed to be created and an exception was thrown, or on connect&query situations we never get access to the connection objects. Though `AsyncMysqlConnectResult` is available to us in most of these situations. Let's implement `sslSessionReused()` in `AsyncMysqlResult`, the base class for `AsyncMysqlConnectResult`, with intention to log its result to `mysql_client_logs` Scuba table, `is_ssl_session_reused` column: https://fburl.com/scuba/mysql_client_logs/eidle5dt

Reviewed By: scotchval

Differential Revision: D35633561

fbshipit-source-id: 672a00ce66c4dd40816d37145dd72f0aa13dc3bd