New bytecode for static method calls with dynamic class name
Summary: We should raise a warning for static method calls with dynamic class names if the class name is a string. Currently the class is already resolved when we execute a `FCallClsMethod`, therefore it cannot differentiate between a string class name and lazy class class name. Create a new bytecode for static method calls with dynamic class names. This bytecode resolves the class name itself and raises a warning if it is a string and appropriate options are set. If the method name is dynamic, we use the existing `FCallClsMethod` bytecode.

Reviewed By: periodic1236

Differential Revision: D35412118

fbshipit-source-id: 3863bae228dedffcebe79cba452c0f0255c3191b