Make all array index read error codes the same
Summary: The type checking code uses different error codes just depending on what code path in the type checker led to the error, even though the errors are really the same. Merge them in preparation for simplifying the array read type checking code, so that it won't have to preserve these arbitrary distinctions.

Reviewed By: enetsee

Differential Revision: D35547742

fbshipit-source-id: 4f52b95ae339ec9b427946f4aac4defbd7e03a06