Fanouts: fix fanout algorithm when modifying a member
Summary:
The current fanout algorithm sometimes fails to recheck member overrides.
This diff fixes it.
When it would fail to recheck: when D::m changes, C <: D and the origin of m in C is D. We would fail to recheck C.
Why it would fail: because there is no edge from C to D::m in this case. This diff fixes that by moving the call to add_idep in Typing_extends further up the call stack.

Reviewed By: bobrenjc93

Differential Revision: D35649755

fbshipit-source-id: 22b2fa370d2d15dc6b8d5ba7f34c2a3755deb6cf