rename all the `fast` stuff to `defs_per_files`
Summary: I went to town and rename anything mentioning a so-called "fast" because that's the most confusing name (fast as speedy? as f-AST?) and we've lost the etymology of it. Instead I went for the more explicit defs_perf_file, since a "fast" was really just a map from file path to top-level definition names.

Reviewed By: bobrenjc93

Differential Revision: D35575539

fbshipit-source-id: 0f9592e772176bae7b72ffca82bdbb77f0c7e1ac