start using new internal field instead of <<__Internal>> attribute in decl parser, nast checks, and typing checks
Summary:
instead of looking for the internal field inside attributes we can
build upon the work in the below diffs and read it from its own top-level
field on classes, function definitions, and type aliases as well as derive it from the
visibility field on method. by removing `has_internal_attribute()` from
naming_attributes_params.ml and compiling we can observe all the callsites to fix -
this diff addresses those.

Reviewed By: jamesjwu

Differential Revision: D35447839

fbshipit-source-id: 40ed069f337b4f1b08d8389f857b36e8f180fcaf