Add new t_internal field to typedef in AST
Summary:
This stack of diffs will tackle 1. adding an "internal" modifier field
to all relevant places in the AST (e.g. functions, classes, traits, interfaces)
on the oCaml side of things 2. updating the Rust view of the AST for parsing
and 3. removing the internal field from file attributes. This is all in the
name of using internal keyword directly instead of as part of an attribute
(e.g. <<__Internal>>)

This diff addresses part 1. for type aliases

Reviewed By: jamesjwu

Differential Revision: D35330707

fbshipit-source-id: e2af6cc6786803cff2280cfbd93990bfd54def95