Push likes without using disjunction
Summary:
The rule for "like-pushing" naturally involves generating a disjunction, as a typical case (for covariant C)
```
C<t> <: ~C<u>
```
can be derived from either
```
t <: u
```
by unions and covariance, or from
```
t <: ~u    /\     u <:D dynamic
```
by covariance, like-pushing, and transitivity.

Our current treatment of disjunctions causes some correct code to be rejected (because we resolve disjunctions early and arbitrarily), and future work on deferring disjunctions is somewhat challenging.

In the common situation that we can detect `u <:D dynamic` at the point that we apply `simplify_subtype`, we can avoid generating a disjunction by simply generating
```
t <: ~u
```
as this subsumes the non-pushing case. For invariant generics, we would generate
```
u <: t <: ~u
```
instead.

The implementation uses an additional flag `super_like` passed around the `simplify_subtype` functions to indicate that the supertype `t` was a component of `~t` and so like pushing is applicable for constructed types such as tuples, shapes, and generic classes.

Differential Revision: D35615041

fbshipit-source-id: 3f483711f923ac64eea8b7a1149e569654d9924f