apc_fetch_with_pure_wakeup, apc_(store|add)_with_pure_sleep
Summary:
I would like apc_fetch to be callable with just read_globals and apc_store/add to be callable with just globals. The former so that it can be called from leak_safe/zoned w/o backdoors and the latter so that they can be called from leak_safe/zoned w/ a special backdoor that does not require resetting the implicit context.

Conceptually APC is just a kind of global state like static properties, except that it can be shared across requests. Generally, we're OK with reading from global state b/c we assume that sensitive data was not put into global state. This is different from reading over arbitrary IO for example, since the kind of information in APC/global state is limited to what is put into it during a request or what the runtime initializes it to be.

Reviewed By: rodmk

Differential Revision: D35565327

fbshipit-source-id: a0edb10f9e0de048ede3c5d32998ce8208e865d4