Add coeffects to gc_collect_cycles
Summary: Garbage collection is internally a complicated process but as far as code is concerned there should be no noticeable difference when this runs (modulo freeing up memory, observable via profiling counters) so `[leak_safe]` feels like a reasonable capability for this. It does affect global state, but not in a directly observable way, so `[globals]` doesn't feel quite right.

Reviewed By: jamesjwu

Differential Revision: D35729498

fbshipit-source-id: 02289fa8f7dcbd5b6b6799f10722c256b76afda6