Skip typechecking bodies of `__Native` functions / methods
Summary:
**Context:** `systemlib` comprises all of the PHP code that is shipped with HHVM and executed but never checked by Hack. This code ought to, and might *have* to, receive the same level of scrutiny we give to any userland piece of Hack code.

The *vast* majority of symbols in systemlib are `__Native`; they are implemented inside the runtime rather than userland. Right now if you ask Hack to typecheck a `__Native` symbol it will assume that it has an empty body. `void` functions and methods are OK, but anything else doesn't work.

On a suggestion from andrewjkennedy, let's type-check these as if they were abstract: we have logic for typechecking methods w/o bodies, let's use it.

Reviewed By: Wilfred

Differential Revision: D35413581

fbshipit-source-id: 1307b482faadc2c9753a8fd06739db6dd7c233c0